Fixed routers
1. Начата работа по упорядочиванию эндпойнтов и контроллеров
