+++ src/app.js
@@ -2,8 +2,9 @@ const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
 
-const servicesRouter = require("./routes/api/servicesRouter");
-const usersRouter = require("./routes/auth/usersRouter");
+const servicesRouter = require("./routes/basic/servicesRouter");
+const authRouter = require("./routes/auth/authRouter");
+const noticesRouter = require("./routes/notices/noticesRouter");
 
 const app = express();
 
@@ -14,7 +15,8 @@ app.use(cors());
 app.use(express.json());
 
 app.use("/", servicesRouter);
-app.use("/auth/users", usersRouter);
+app.use("/auth", authRouter);
+app.use("/notices", noticesRouter);
 
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
