+++ src/middlewares/authMiddleware.js
@@ -2,6 +2,7 @@ const jwt = require("jsonwebtoken");
 const { User } = require("../schemas/user");
 const { ACCESS_SECRET_KEY } = process.env;
 
+// Написати прошарок авторизації
 const authMiddleware = async (req, res, next) => {
   const { authorization = "" } = req.headers;
   const [bearer, token] = authorization.split(" ");
@@ -13,7 +14,7 @@ const authMiddleware = async (req, res, next) => {
           "Please, provide a token in request authorization header",
       });
     }
-    
+
     const { id } = jwt.verify(token, ACCESS_SECRET_KEY);
     if (!id) {
       return res.status(401).json({
